+++ Selu383.SP25.P03.Web/src/pages/leases-page/EditLeases.tsx
@@ -220,14 +220,20 @@ const EditLeases: React.FC<EditLeasesProps> = ({ currentUser }) => {
 
   return (
     <div className="p-20px max-w-1200px mx-auto">
-      <h1 className="text-gray-800">Manage Leases</h1>
-      <p className="text-gray-700">Logged in as: {currentUser.userName} (ID: {currentUser.id})</p>
+      <h1 className="text-gray-800 text-2xl font-semibold mb-10px">Manage Leases</h1>
       
-      <form onSubmit={handleSubmit} className="bg-[#00061f] text-white p-20px rounded-8px mb-30px">
-        <h2>{editingId ? "Edit Lease" : "Add New Lease"}</h2>
-        
-        <div className="mb-15px">
-          <label htmlFor="tenantId" className="block mb-5px font-bold">Tenant:</label>
+      <form 
+        onSubmit={handleSubmit} 
+        className="bg-white text-gray-800 shadow-lg p-24px rounded-12px border border-gray-300 mb-30px"
+      >
+        <h2 className="text-lg font-semibold mb-24px">
+          {editingId ? "Edit Lease" : "Create Lease"}
+        </h2>
+
+        <div className="mb-20px">
+          <label htmlFor="tenantId" className="block mb-6px font-medium text-gray-700">
+            Tenant
+          </label>
           <select
             id="tenantId"
             name="tenantId"
@@ -244,7 +250,7 @@ const EditLeases: React.FC<EditLeasesProps> = ({ currentUser }) => {
                 lastName: tenant?.lastName || ""
               }));
             }}
-            className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
+            className="w-full px-3 py-2 border border-gray-300 rounded-lg shadow-inner bg-white focus:(outline-none ring-2 ring-blue-400)"
             required
           >
             <option value={0}>-- Select Tenant --</option>
@@ -254,119 +260,139 @@ const EditLeases: React.FC<EditLeasesProps> = ({ currentUser }) => {
               </option>
             ))}
           </select>
-          {userTenants.length === 0 && (
-            <p className="text-[#dc3545] text-12px mt-5px">No tenants found for your properties.</p>
-          )}
         </div>
 
-        <div className="mb-15px">
-            <label htmlFor="startDate" className="block mb-5px font-bold">Start Date:</label>
+        <div className="grid grid-cols-1 md:grid-cols-2 mb-5">
+          <div className="flex flex-col">
+            <label
+              htmlFor="startDate"
+              className="mb-2 font-medium text-gray-700"
+            >
+              Start Date
+            </label>
             <input
-                type="date"
-                id="startDate"
-                name="startDate"
-                value={formData.startDate}
-                onChange={handleInputChange}
-                className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
-                required
+              type="date"
+              id="startDate"
+              name="startDate"
+              value={formData.startDate}
+              onChange={handleInputChange}
+              className="w-100 px-3 py-2 border border-gray-300 rounded-lg shadow-inner bg-white focus:(outline-none ring-2 ring-blue-400)"
+              required
             />
-        </div>
+          </div>
 
-        <div className="mb-15px">
-            <label htmlFor="endDate" className="block mb-5px font-bold">End Date:</label>
+          <div className="flex flex-col">
+            <label
+              htmlFor="endDate"
+              className="mb-2 font-medium text-gray-700"
+            >
+              End Date
+            </label>
             <input
-                type="date"
-                id="endDate"
-                name="endDate"
-                value={formData.endDate}
-                onChange={handleInputChange}
-                className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
-                required
+              type="date"
+              id="endDate"
+              name="endDate"
+              value={formData.endDate}
+              onChange={handleInputChange}
+              className="w-105 px-3 py-2 border border-gray-300 rounded-lg shadow-inner focus:(outline-none ring-2 ring-blue-400)"
+              required
             />
+          </div>
+        </div>
+        {/* Rent */}
+        <div className="mb-20px">
+          <label htmlFor="rent" className="block mb-6px font-medium text-gray-700">
+            Rent
+          </label>
+          <div className="flex items-center gap-8px">
+            <span className="text-gray-500">$</span>
+            <input
+              type="number"
+              id="rent"
+              name="rent"
+              value={formData.rent === 0 ? "" : formData.rent}
+              onChange={handleInputChange}
+              className="flex-1 px-12px py-10px border border-gray-300 rounded-8px text-14px focus:(outline-none ring-2 ring-blue-400)"
+              required
+            />
+          </div>
         </div>
 
-        <div className="mb-15px">
-            <label htmlFor="rent" className="block mb-5px font-bold">Rent:</label>
-            <div className="flex items-center gap-5px">
-                <span>$</span>
-                <input
-                    type="number"
-                    id="rent"
-                    name="rent"
-                    value={formData.rent === 0 ? "" : formData.rent}
-                    onChange={handleInputChange}
-                    className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
-                    required
-                />
-            </div>
+        {/* Deposit */}
+        <div className="mb-20px">
+          <label htmlFor="deposit" className="block mb-6px font-medium text-gray-700">
+            Deposit
+          </label>
+          <div className="flex items-center gap-8px">
+            <span className="text-gray-500">$</span>
+            <input
+              type="number"
+              id="deposit"
+              name="deposit"
+              value={formData.deposit === 0 ? "" : formData.deposit}
+              onChange={handleInputChange}
+              className="flex-1 px-12px py-10px border border-gray-300 rounded-8px text-14px focus:(outline-none ring-2 ring-blue-400)"
+              required
+            />
+          </div>
         </div>
 
-        <div className="mb-15px">
-            <label htmlFor="deposit" className="block mb-5px font-bold">Deposit:</label>
-            <div className="flex items-center gap-5px">
-                <span>$</span>
-                <input
-                    type="number"
-                    id="deposit"
-                    name="deposit"
-                    value={formData.deposit === 0 ? "" : formData.deposit}
-                    onChange={handleInputChange}
-                    className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
-                    required
-                />
-            </div>
-        </div>
-
-        <div className="mb-15px">
-            <label htmlFor="status" className="block mb-5px font-bold">Status:</label>
-            <select
-                id="status"
-                name="status"
-                value={formData.status}
-                onChange={handleInputChange}
-                className="w-full p-8px border-1 border-[#ddd] rounded-4px text-14px"
-                required
-            >     
-                <option value="">-- Select Status --</option>
-                <option value="Active">Active</option>
-                <option value="Terminated">Terminated</option>
-                <option value="Expired">Expired</option>
-            </select>
+        {/* Status */}
+        <div className="mb-20px">
+          <label htmlFor="status" className="block mb-6px font-medium text-gray-700">
+            Status
+          </label>
+          <select
+            id="status"
+            name="status"
+            value={formData.status}
+            onChange={handleInputChange}
+            className="w-full px-12px py-10px border border-gray-300 rounded-8px text-14px focus:(outline-none ring-2 ring-blue-400)"
+            required
+          >
+            <option value="">-- Select Status --</option>
+            <option value="Active">Active</option>
+            <option value="Terminated">Terminated</option>
+            <option value="Expired">Expired</option>
+          </select>
         </div>
 
         {error && showMessage && (
-          <div className="text-[#721c24] my-10px py-10px px-15px bg-[#f8d7da] border border-[#f5c6cb] rounded-4px">
+          <div className="my-4 p-4 rounded-lg shadow-inner border border-red-300 bg-red-100 text-red-800 text-sm">
             {error}
           </div>
         )}
 
         {!error && message && (
-          <div className="text-[#155724] my-10px py-10px px-15px bg-[#d4edda] border border-[#c3e6cb] rounded-4px">
+          <div className="my-4 p-4 rounded-lg shadow-inner border border-green-300 bg-green-100 text-green-800 text-sm">
             {message}
           </div>
         )}
 
-        <div className="flex gap-10px mt-20px">
+        {/* Buttons */}
+        <div className="flex flex-wrap gap-12px mt-24px">
           <button 
             type="submit" 
             disabled={loading || userTenants.length === 0}
-            className="bg-[#007bff] text-white py-10px px-20px border-none rounded-4px cursor-pointer text-14px hover:bg-[#0056b3] disabled:opacity-60 disabled:cursor-not-allowed"
+            className="bg-blue-500 text-white py-10px px-20px rounded-8px text-14px hover:bg-blue-700 disabled:opacity-60 disabled:cursor-not-allowed transition-colors"
           >
             {loading ? "Saving..." : editingId ? "Update Lease" : "Add Lease"}
           </button>
+
           {editingId && (
             <button 
               type="button" 
-              onClick={resetForm} 
+              onClick={resetForm}
               disabled={loading}
-              className="bg-[#6c757d] text-white py-10px px-20px border-none rounded-4px cursor-pointer text-14px hover:bg-[#545b62] disabled:opacity-60 disabled:cursor-not-allowed"
+              className="bg-gray-400 text-white py-10px px-20px rounded-8px text-14px hover:bg-gray-500 disabled:opacity-60 disabled:cursor-not-allowed transition-colors"
             >
               Cancel
             </button>
           )}
+
           <Link 
             to="/leases"
-            className="bg-green-600 text-white py-2 px-4 rounded-md text-sm hover:bg-green-700 transition-colors duration-200"
+            className= "bg-blue-500 text-white py-10px px-20px rounded-8px text-14px hover:bg-blue-700"
           >
             View All Leases
           </Link>
@@ -374,6 +400,7 @@ const EditLeases: React.FC<EditLeasesProps> = ({ currentUser }) => {
       </form>
     </div>
   );
+  
 };
 
 export default EditLeases;
