+++ Selu383.SP25.P03.Web/src/pages/tenants-page/ListTenants.tsx
@@ -40,7 +40,7 @@ const ListTenants: React.FC<ListTenantsProps> = ({ currentUser }) => {
   const [message, setMessage] = useState("");
   const [showMessage, setShowMessage] = useState(false);
   const [searchTerm, setSearchTerm] = useState(""); 
-  const [searchField, setSearchField] = useState<"firstName" | "lastName" | "unitNumber">("firstName"); // 🔹 new state
+  const [searchField, setSearchField] = useState<"firstName" | "lastName" | "unitNumber">("firstName");
 
   useEffect(() => {
       if (currentUser) {
@@ -122,81 +122,104 @@ const ListTenants: React.FC<ListTenantsProps> = ({ currentUser }) => {
 
   return (
       <div className="tenants-list">
-        <h2>Your Tenants ({filteredTenants.length})</h2>
-        <div className="flex gap-2 mb-4">
-            <select
+        <h2>Tenants ({filteredTenants.length})</h2>
+        <div className="flex items-center w-full max-w-xl gap-2">
+          {/* Search Field Dropdown */}
+          <select
             value={searchField}
-            onChange={(e) => setSearchField(e.target.value as "firstName" | "lastName" | "unitNumber")}
-            className="px-3 py-2 rounded border text-black"
-            >
+            onChange={(e) =>
+              setSearchField(
+                e.target.value as "firstName" | "lastName" | "unitNumber"
+              )
+            }
+            className="px-3 py-2 rounded-full border border-gray-300 text-black bg-white shadow-sm focus:ring-2 focus:ring-blue-500 focus:outline-none"
+          >
             <option value="firstName">First Name</option>
             <option value="lastName">Last Name</option>
-            <option value="unitNumber">Unit Number</option>
-            </select>
+            <option value="unitNumber">Unit</option>
+          </select>
 
+          {/* Search Input */}
+          <div className="relative flex-1">
+            <span className="absolute inset-y-0 left-3 flex items-center text-gray-400">
+              <svg
+                className="w-5 h-5 text-gray-400"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth={2}
+                viewBox="0 0 24 24"
+              >
+                <circle cx="11" cy="11" r="8" />
+                <line x1="21" y1="21" x2="16.65" y2="16.65" />
+              </svg>
+            </span>
             <input
-            type="text"
-            placeholder={`Search by ${searchField}...`}
-            value={searchTerm}
-            onChange={(e) => setSearchTerm(e.target.value)}
-            className="px-3 py-2 rounded border text-black flex-1"
+              type="text"
+              placeholder={`Search by ${searchField}...`}
+              value={searchTerm}
+              onChange={(e) => setSearchTerm(e.target.value)}
+              className="pl-10 pr-4 py-2 w-full rounded-full border border-gray-300 focus:ring-2 focus:ring-blue-500 focus:outline-none text-black shadow-sm"
             />
+          </div>
         </div>
         {filteredTenants.length === 0 ? (
           <p>No Tenants Found</p>
         ) : (
 
            
-          <table className="w-full text-white border-collapse mt-20px">
+          <table className="w-full border-collapse mt-5 text-sm bg-[#fdfefe] rounded-lg overflow-hidden shadow-sm">
             <thead>
-              <tr>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Unit</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">First Name</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Last Name</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Phone Number</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Email</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Created</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Updated</th>
-                <th className="p-12px text-left border-b-1 border-[#ddd] bg-[#01101f] font-bold">Actions</th>
+              <tr className="bg-[#f3f4f6]">
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Unit</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">First Name</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Last Name</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Phone Number</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Email</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Created</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Updated</th>
+                <th className="p-12px text-left border-b border-r border-[#e5e7eb] font-semibold text-[#374151]">Actions</th>
               </tr>
             </thead>
             <tbody>
-              {filteredTenants.map(t => (
-                <tr key={t.id} className="bg-[#322c35]">
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{t.unitNumber}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{t.firstName}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{t.lastName}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{t.phoneNumber}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{t.email}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{new Date(t.createdAt).toLocaleDateString()}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">{new Date(t.updatedAt).toLocaleDateString()}</td>
-                  <td className="p-12px text-left border-b-1 border-[#ddd]">
+              {filteredTenants.map((t,i) => (
+                <tr
+                  key={t.id}
+                  className={`${i % 2 === 0 ? "bg-white" : "bg-[#f9fafb]"} hover:bg-[#ebf5ff] transition-colors`}
+                >
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{t.unitNumber}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{t.firstName}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{t.lastName}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{t.phoneNumber}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{t.email}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{new Date(t.createdAt).toLocaleDateString()}</td>
+                  <td className="p-12px border-b border-r border-[#e5e7eb] text-[#111827]">{new Date(t.updatedAt).toLocaleDateString()}</td>
+                  <td className="p-12px border-b flex gap-2">
                     <Link
-                        to={`/edittenants/${t.id}`}
-                        className="bg-[#28a745] text-white py-6px px-12px mr-5px border-none rounded-4px cursor-pointer text-12px hover:bg-[#1e7e34]"
+                      to={`/edittenants/${t.id}`}
+                      className="bg-[#22c55e] text-white py-6px px-12px rounded-md text-12px hover:bg-[#1e7e34] transition-colors"
                     >
-                        Edit
+                      Edit
                     </Link>
                     <button
                       onClick={() => t.id && handleDelete(t.id)}
-                      className="bg-[#dc3545] text-white py-6px px-12px border-none rounded-4px cursor-pointer text-12px hover:bg-[#c82333] disabled:opacity-60 disabled:cursor-not-allowed"
+                      className="bg-[#ef4444] text-white py-6px px-12px rounded-md text-12px hover:bg-[#dc2626] transition-colors disabled:opacity-60 disabled:cursor-not-allowed"
                     >
-                      Delete
+                     Delete
                     </button>
                   </td>
                 </tr>
               ))}
             </tbody>
             {error && showMessage && (
-            <div className="text-[#721c24] my-10px py-10px px-15px bg-[#f8d7da] border border-[#f5c6cb] rounded-4px">
+              <div className="my-4 p-4 rounded-lg shadow-inner border border-red-300 bg-red-100 text-red-800 text-sm">
                 {error}
-            </div>
+              </div>
             )}
 
             {!error && message && (
-            <div className="text-[#155724] my-10px py-10px px-15px bg-[#d4edda] border border-[#c3e6cb] rounded-4px">
+              <div className="my-4 p-4 rounded-lg shadow-inner border border-green-300 bg-green-100 text-green-800 text-sm">
                 {message}
-            </div>
+              </div>
             )}
           </table>
         )}
